+++ src/components/atom/QuestionBox/index.tsx
@@ -7,6 +7,7 @@ import { MainSemiTitle } from "../MainSemiTitle";
 import { PublicButton } from "../button";
 import { Container, PreviewContent } from "./style";
 import { useMarkdown } from "@/hooks/useMarkdown";
+import { handleMarkdownKeyDown } from "@/utils/markdownKeyHandlers";
 
 // 툴바 타입 명확화
 interface ToolbarItem {
@@ -97,38 +98,19 @@ export const QuestionBox = ({ data, idx }: { data: BasicQuestionType; idx: numbe
   // handleKeyDown (자동 리스트)
   const handleKeyDown = useCallback(
     (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      if (isComposing) return;
-      if (e.key === "Enter") {
-        const textarea = e.currentTarget;
-        const start = textarea.selectionStart;
-        const end = textarea.selectionEnd;
-        const value = text;
-        let lineStart = start;
-        while (lineStart > 0 && value[lineStart - 1] !== "\n") {
-          lineStart--;
-        }
-        const currentLine = value.substring(lineStart, start);
-        const listMatch = currentLine.match(/^(\s*)([-*+]|\d+\.)\s(.*)$/);
-        if (listMatch) {
-          e.preventDefault();
-          const [_, indent, marker, content] = listMatch;
-          const before = value.substring(0, start);
-          const after = value.substring(end);
-          let newValue: string;
-          if (content.trim() === "") {
-            newValue = before + "\n" + after;
-          } else {
-            const newLine = "\n" + indent + marker + " ";
-            newValue = before + newLine + after;
-          }
-          setText(newValue);
+      handleMarkdownKeyDown({
+        e,
+        isComposing,
+        text,
+        setText,
+        updateRecoil: (newValue) => {
           setpostWritingData(
             postWritingData.map((item) =>
               item.questionId === data.questionId ? { ...item, content: newValue } : item
             )
           );
-        }
-      }
+        },
+      });
     },
     [isComposing, text, setText, setpostWritingData, postWritingData, data.questionId]
   );
